add theme toggle and split screen layout
