+++ src/components/Header.tsx
@@ -1,4 +1,6 @@
-import { config, ids } from '@/data/config';
+import photo from '@/assets/kelly-ripple-circle.png';
+import { SocialIcon } from '@/components/SocialIcon';
+import { config } from '@/data/config';
 import { sections } from '@/data/sections';
 
 import '@/components/header.css';
@@ -6,39 +8,42 @@ import '@/components/header.css';
 export function Header({ children }: { children: ReactNode }) {
   return (
     <header className="header">
-      <div className="app-controls">
-        <a
-          className="app-title click-to-close-menu click-to-scroll"
-          href="#home"
-        >
-          {config.name}
-        </a>
-        <label
-          aria-label="Menu button"
-          className="menu-button"
-          htmlFor={ids.menuButton}
-        ></label>
+      <div>
+        <div className="header-image-container">
+          <div className="header-image-frame">
+            <img
+              alt={config.name}
+              className="header-image-circle"
+              src={photo}
+            ></img>
+          </div>
+        </div>
+        <h1 className="name">{config.name}</h1>
+        <h2 className="job-title">{config.jobTitle}</h2>
+        <div className="header-actions">
+          <nav className="nav" role="navigation">
+            {Object.values(sections).map((section) => {
+              return (
+                <a
+                  className="nav-link click-to-scroll"
+                  href={`#${section.id}`}
+                  key={section.id}
+                >
+                  <span className="nav-indicator"></span>
+                  {section.id}
+                </a>
+              );
+            })}
+          </nav>
+          {children}
+        </div>
       </div>
-      <nav className="nav" role="navigation">
-        <a
-          className="section-link click-to-close-menu click-to-scroll"
-          href={`#${ids.home}`}
-        >
-          {ids.home}
-        </a>
-        {Object.keys(sections).map((id) => {
-          return (
-            <a
-              className="section-link click-to-close-menu click-to-scroll"
-              href={`#${id}`}
-              key={id}
-            >
-              {id}
-            </a>
-          );
-        })}
-        {children}
-      </nav>
+
+      <div className="icon-set">
+        <SocialIcon icon="github" />
+        <SocialIcon icon="linkedin" />
+        <SocialIcon icon="email" />
+      </div>
     </header>
   );
 }
