+++ src/components/Icon.tsx
@@ -1,30 +1,50 @@
-import { CopyButton } from '@/components/CopyButton';
-import { SvgIcon } from '@/components/SvgIcon';
-import { config } from '@/data/config';
-
 import '@/components/icon.css';
 
-const github = `https://github.com/${config.github}` as const;
-const linkedin = `https://www.linkedin.com/in/${config.linkedin}` as const;
-
 const icons = {
+  api: {
+    viewBox: '1.83 1.83 20.33 20.33',
+    drawPath:
+      'M13 13c-.56.56-1.45.56-2 .01V13c-.55-.55-.55-1.44 0-1.99V11c.55-.55 1.44-.55 1.99 0H13c.55.55.55 1.45 0 2m-1-7 2.12 2.12 2.5-2.5-3.2-3.2c-.78-.78-2.05-.78-2.83 0l-3.2 3.2 2.5 2.5zm-6 6 2.12-2.12-2.5-2.5-3.2 3.2c-.78.78-.78 2.05 0 2.83l3.2 3.2 2.5-2.5zm12 0-2.12 2.12 2.5 2.5 3.2-3.2c.78-.78.78-2.05 0-2.83l-3.2-3.2-2.5 2.5zm-6 6-2.12-2.12-2.5 2.5 3.2 3.2c.78.78 2.05.78 2.83 0l3.2-3.2-2.5-2.5z',
+  },
+  copy: {
+    viewBox: '2 2 20 20',
+    drawPath:
+      'M7 5a3 3 0 0 1 3-3h9a3 3 0 0 1 3 3v9a3 3 0 0 1-3 3h-2v2a3 3 0 0 1-3 3H5a3 3 0 0 1-3-3v-9a3 3 0 0 1 3-3h2V5Zm2 2h5a3 3 0 0 1 3 3v5h2a1 1 0 0 0 1-1V5a1 1 0 0 0-1-1h-9a1 1 0 0 0-1 1v2ZM5 9a1 1 0 0 0-1 1v9a1 1 0 0 0 1 1h9a1 1 0 0 0 1-1v-9a1 1 0 0 0-1-1H5Z',
+  },
   email: {
-    href: `mailto:${config.email}`,
-    label: config.email,
-    title: 'Email',
-    value: config.email,
+    viewBox: '-1081.46 -287.22 21.5 21.5',
+    drawPath:
+      'M-1070.71-287.216a10.75 10.75 0 0 0-10.75 10.75 10.75 10.75 0 0 0 10.75 10.75 10.75 10.75 0 0 0 10.75-10.75 10.75 10.75 0 0 0-10.75-10.75zm-6.074 5.803h12.15l-6.076 4.338zm-.852.608 6.64 4.74a.495.495 0 0 0 .574 0l6.64-4.74v8.791a.495.495 0 0 1-.496.496h-12.862a.495.495 0 0 1-.496-.496z',
   },
   github: {
-    href: github,
-    label: `@${config.github}`,
-    title: 'GitHub',
-    value: github,
+    viewBox: '0.92 1.27 22.15 21.47',
+    drawPath:
+      'M12 1.27a11 11 0 00-3.48 21.46c.55.09.73-.28.73-.55v-1.84c-3.03.64-3.67-1.46-3.67-1.46-.55-1.29-1.28-1.65-1.28-1.65-.92-.65.1-.65.1-.65 1.1 0 1.73 1.1 1.73 1.1.92 1.65 2.57 1.2 3.21.92a2 2 0 01.64-1.47c-2.47-.27-5.04-1.19-5.04-5.5 0-1.1.46-2.1 1.2-2.84a3.76 3.76 0 010-2.93s.91-.28 3.11 1.1c1.8-.49 3.7-.49 5.5 0 2.1-1.38 3.02-1.1 3.02-1.1a3.76 3.76 0 010 2.93c.83.74 1.2 1.74 1.2 2.94 0 4.21-2.57 5.13-5.04 5.4.45.37.82.92.82 2.02v3.03c0 .27.1.64.73.55A11 11 0 0012 1.27',
+  },
+  internet: {
+    viewBox: '85.33 85.33 853.33 853.33',
+    drawPath:
+      'M698.027 597.333C701.44 569.173 704 541.013 704 512c0-29.013-2.56-57.173-5.973-85.333H842.24c6.827 27.306 11.093 55.893 11.093 85.333 0 29.44-4.266 58.027-11.093 85.333M622.507 834.56c25.6-47.36 45.226-98.56 58.88-151.893h125.866c-40.96 70.4-106.24 125.013-184.746 151.893M611.84 597.333H412.16c-4.267-28.16-6.827-56.32-6.827-85.333 0-29.013 2.56-57.6 6.827-85.333h199.68c3.84 27.733 6.827 56.32 6.827 85.333 0 29.013-2.987 57.173-6.827 85.333M512 851.627c-35.413-51.2-64-107.947-81.493-168.96h162.986C576 743.68 547.413 800.427 512 851.627M341.333 341.333H216.747c40.533-70.826 106.24-125.44 184.32-151.893-25.6 47.36-44.8 98.56-59.734 151.893M216.747 682.667h124.586C356.267 736 375.467 787.2 401.067 834.56c-78.08-26.88-143.787-81.493-184.32-151.893m-34.987-85.334c-6.827-27.306-11.093-55.893-11.093-85.333 0-29.44 4.266-58.027 11.093-85.333h144.213C322.56 454.827 320 482.987 320 512c0 29.013 2.56 57.173 5.973 85.333M512 171.947c35.413 51.2 64 108.373 81.493 169.386H430.507C448 280.32 476.587 223.147 512 171.947m295.253 169.386H681.387C667.733 288 648.107 236.8 622.507 189.44c78.506 26.88 143.786 81.067 184.746 151.893M512 85.333c-235.947 0-426.667 192-426.667 426.667 0 235.52 191.147 426.667 426.667 426.667 235.52 0 426.667-191.147 426.667-426.667C938.667 276.48 747.52 85.333 512 85.333z',
   },
   linkedin: {
-    href: linkedin,
-    label: `/in/${config.linkedin}`,
-    title: 'LinkedIn',
-    value: linkedin,
+    viewBox: '-426.41 617.69 21.5 21.5',
+    drawPath:
+      'M-415.663 617.69a10.75 10.75 0 0 0-10.75 10.75 10.75 10.75 0 0 0 10.75 10.75 10.75 10.75 0 0 0 10.75-10.75 10.75 10.75 0 0 0-10.75-10.75zm-4.965 4.094c.93 0 1.68.761 1.68 1.691a1.68 1.68 0 0 1-1.68 1.68c-.93 0-1.691-.75-1.691-1.68a1.69 1.69 0 0 1 1.691-1.691zm8.36 4.75a3.26 3.26 0 0 1 3.26 3.26v5.3h-2.79v-4.93a1.4 1.4 0 0 0-1.4-1.4c-.77 0-1.39.63-1.39 1.4v4.93h-2.79v-8.37h2.79v1.11c.48-.78 1.47-1.3 2.32-1.3zm-9.74.19h2.769v8.37h-2.77z',
+  },
+  moon: {
+    viewBox: '0 0 24 24',
+    drawPath:
+      'M12.0105 0A12.0105 12.0105 0 0 0 0 12.0105a12.0105 12.0105 0 0 0 12.0105 12.0105 12.0105 12.0105 0 0 0 12.0105 -12.0105A12.0105 12.0105 0 0 0 23.455 8.38428a6.91703 6.91703 0 0 1 -6.81223 5.79563 6.91703 6.91703 0 0 1 -6.91703 -6.91703 6.91703 6.91703 0 0 1 5.65939 -6.79127A12.0105 12.0105 0 0 0 12 0z',
+  },
+  plus: {
+    viewBox: '5 5 14 14',
+    drawPath:
+      'M18 13h-5v5c0 .55-.45 1-1 1s-1-.45-1-1v-5H6c-.55 0-1-.45-1-1s.45-1 1-1h5V6c0-.55.45-1 1-1s1 .45 1 1v5h5c.55 0 1 .45 1 1s-.45 1-1 1',
+  },
+  sun: {
+    viewBox: '0 0 24 24',
+    drawPath:
+      'M11.96.079c-.406.016-.73.38-.73.826v2.188c0 .461.34.834.77.834.421 0 .77-.373.77-.834V.905c0-.462-.349-.826-.77-.826h-.049zm8.004 3.216a.867.867 0 0 0-.616.26l-1.54 1.539c-.331.332-.356.842-.048 1.134.3.308.81.292 1.134-.04l1.54-1.54c.324-.324.348-.834.048-1.134a.737.737 0 0 0-.518-.219zm-15.88.065a.737.737 0 0 0-.51.219c-.307.291-.291.802.033 1.126l1.54 1.54c.331.323.842.347 1.133.048.324-.3.308-.81-.024-1.135L4.709 3.603a.859.859 0 0 0-.616-.251zM12 7.767a4.229 4.229 0 0 0-4.237 4.237A4.229 4.229 0 0 0 12 16.234a4.229 4.229 0 0 0 4.229-4.23A4.229 4.229 0 0 0 12 7.767zM.836 11.235c-.462 0-.834.34-.834.77s.372.769.834.769h2.188c.461 0 .834-.34.834-.77s-.373-.786-.834-.786zm20.14 0c-.461 0-.834.34-.834.77s.373.769.834.769h2.188c.462 0 .834-.34.834-.77s-.372-.77-.834-.77zm-2.795 6.254a.737.737 0 0 0-.486.219c-.3.3-.275.81.049 1.134l1.547 1.54c.324.323.835.348 1.135.048.3-.3.283-.81-.041-1.134l-1.547-1.54a.859.859 0 0 0-.649-.267zm-12.5.049a.867.867 0 0 0-.575.259l-1.54 1.547c-.332.324-.356.835-.048 1.134.3.3.81.284 1.134-.048l1.54-1.54c.323-.324.348-.834.048-1.134a.745.745 0 0 0-.56-.218zm6.278 2.527c-.405.016-.729.381-.729.827v2.187c0 .462.34.834.77.834.421 0 .77-.372.77-.834v-2.187c0-.462-.349-.827-.77-.827h-.049z',
   },
 } as const;
 
@@ -32,21 +52,8 @@ export function Icon({ icon: key }: { icon: keyof typeof icons }) {
   const icon = icons[key];
 
   return (
-    <div className="social-icons">
-      <a
-        aria-label={key}
-        className="icon-link"
-        href={icon.href}
-        rel="noreferrer"
-        target="_blank"
-      >
-        <SvgIcon icon={key} />
-        <span className="icon-labels">
-          <span className="icon-label-title">{icon.title}</span>
-          <span className="icon-label-details">{icon.label}</span>
-        </span>
-      </a>
-      <CopyButton value={icon.value} />
-    </div>
+    <svg className={`icon ${key}-icon`} viewBox={icon.viewBox}>
+      <path d={icon.drawPath} />
+    </svg>
   );
 }
