--- src/components/Image.tsx
@@ -1,9 +1,0 @@
-import '@/components/image.css';
-
-export function Image(props: { alt: string; src: string }) {
-  return (
-    <div className="image-container">
-      <img className="image" {...props}></img>
-    </div>
-  );
-}
