+++ src/components/icon.css
@@ -1,30 +1,6 @@
-.social-icons {
-  align-items: center;
-  display: flex;
-}
-
-.social-icons .icon-link {
-  align-items: center;
-  display: flex;
-  flex-grow: 1;
-  flex-shrink: 0;
-  gap: var(--icon-label-gap);
-  padding: var(--icon-link-padding);
-  text-decoration: none;
-}
-
-.social-icons .icon-labels {
-  display: flex;
-  flex-direction: column;
-  gap: 2px;
-}
-
-.social-icons .icon-label-title {
-  font-size: 1rem;
-}
-
-.social-icons .icon-label-details {
-  font-size: var(--padding-half);
-  font-weight: 300;
-  opacity: 0.75;
+.icon {
+  fill: currentColor;
+  height: var(--icon-size);
+  padding: var(--icon-padding);
+  width: var(--icon-size);
 }
