--- src/components/image.css
@@ -1,29 +1,0 @@
-.image-container {
-  align-items: center;
-  border-radius: 50%;
-  display: flex;
-  flex-shrink: 0;
-  justify-content: center;
-  margin: 16px;
-  position: relative;
-  width: 55%;
-}
-
-.image-container::after {
-  border-radius: inherit;
-  bottom: 0;
-  box-shadow: inset 0 0 8px 8px var(--theme-color);
-  content: '';
-  left: 0;
-  position: absolute;
-  right: 0;
-  top: 0;
-}
-
-.image {
-  /* background-color: white;
-  border: 2px solid white; */
-  border-radius: inherit;
-  height: auto;
-  width: 100%;
-}
