--- src/components/svg-icon.css
@@ -1,6 +1,0 @@
-.icon {
-  fill: currentColor;
-  height: var(--icon-size);
-  padding: var(--icon-padding);
-  width: var(--icon-size);
-}
