+++ src/components/theme.css
@@ -3,29 +3,67 @@
   --eerie-black-overlay: #222222cc;
   --platinum: #dddddd;
   --platinum-overlay: #ddddddcc;
-
+}
+:root {
   --oxford-blue: #04052e;
   --oxford-blue-overlay: #04052ecc;
   --dark-purple: #2f1b25;
   --carmine: #950017;
   --chocolate-cosmos: #550022;
-
+}
+:root {
   --powder-blue: #afc1d6;
   --powder-blue-overlay: #afc1d6cc;
   --african-violet: #ab92bf;
   --mint-green: #cef9f2;
   --ultra-violet: #655a7c;
+}
 
-  /* these are used to set the html background-color */
-  --theme-color-dark: var(--oxford-blue);
-  --theme-color-light: var(--powder-blue);
+:root {
+  --test: rgb(253, 0, 115);
+  --test2: rgb(19, 22, 82);
+  --test5: rgb(38, 27, 98);
+  --test3: rgb(252, 255, 255);
+  --test4: rgb(254, 254, 254);
+  --test4: rgb(123, 123, 123);
+  --test6: rgb(23, 9, 55);
+}
+
+:root {
+  --coquelicot: #ff4000;
+  --scarlet: #ff2a00;
+  --folly: #ff0044;
+  --rose: #ff006f;
+  --magenta: #ff0088;
+  --electric-purple: #b300ff;
+  --violet: #8400ff;
+  --palatinate-blue: #3158f8;
+  --risd-blue: #0059ff;
+  --celestial-blue: #009dff;
+  --electric-blue: #00eaff;
+  --aquamarine: #00ffd5;
+  --white: #ffffff;
+}
+
+/* these are used to set the html background-color */
+:root {
+  /* --theme-color-dark: var(--oxford-blue);
+  --theme-color-light: var(--powder-blue); */
+  --theme-color-dark: var(--dark-purple);
+  --theme-color-light: var(--african-violet);
 }
 
 /* dark mode (default) */
 :root {
   --theme-color: var(--theme-color-dark);
   --text-color: var(--platinum);
+  --text-color-emphasis: #ffffff;
   --contrast-color: var(--eerie-black);
+
+  /* --theme-toggle-color: #229fea;
+  --theme-toggle-background-color: #0b2231;
+  --theme-toggle-background-color: #229fea99; */
+
   --opacity-dark: 1;
   --opacity-light: 0;
 
@@ -35,14 +73,26 @@
 
   --theme-color-overlay: var(--oxford-blue-overlay);
   --text-color-overlay: var(--platinum-overlay);
-  --highlight-color: cyan;
-  --backdrop-color: transparent;
 
-  --tag-color: rgba(0, 230, 230, 0.1);
+  --tag-backdrop-color: transparent;
+  --backdrop-color: #ffffff66;
+
+  /* --aquamarine */
   --tag-text-rgb: 0, 255, 213;
+  --tag-color: rgba(0, 230, 230, 0.1);
   --tag-text-color: rgb(var(--tag-text-rgb));
 
+  --highlight-color: rgba(var(--tag-text-rgb), 0.9);
+  --highlight-text-color: #000000;
+
   --about-section-contrast-color: rgba(var(--tag-text-rgb), 0.3);
+
+  --gradient: linear-gradient(
+    123deg,
+    var(--coquelicot) 0%,
+    var(--violet) 50%,
+    var(--palatinate-blue) 100%
+  );
 }
 
 /* light mode */
@@ -50,7 +100,13 @@ html[data-theme='light'],
 #theme-toggle:checked ~ .theme {
   --theme-color: var(--theme-color-light);
   --text-color: var(--eerie-black);
+  --text-color-emphasis: #000000;
   --contrast-color: var(--platinum);
+
+  /* --theme-toggle-color: #ffffff;
+  --theme-toggle-background-color: #f8e6a1;
+  --theme-toggle-background-color: #86671988; */
+
   --opacity-dark: 0;
   --opacity-light: 1;
 
@@ -60,14 +116,26 @@ html[data-theme='light'],
 
   --theme-color-overlay: var(--powder-blue-overlay);
   --text-color-overlay: var(--eerie-black-overlay);
-  --highlight-color: cyan;
-  --backdrop-color: #ffffff66;
 
-  --tag-color: rgba(0, 21, 255, 0.1);
+  --tag-backdrop-color: #ffffff66;
+  --backdrop-color: #ffffff99;
+
+  /* --risd-blue */
   --tag-text-rgb: 0, 89, 255;
+  --tag-color: rgba(0, 21, 255, 0.1);
   --tag-text-color: rgb(var(--tag-text-rgb));
 
+  --highlight-color: rgba(var(--tag-text-rgb), 0.8);
+  --highlight-text-color: #ffffff;
+
   --about-section-contrast-color: var(--platinum);
+
+  --gradient: linear-gradient(
+    123deg,
+    var(--scarlet) 0%,
+    var(--electric-purple) 50%,
+    var(--risd-blue) 100%
+  );
 }
 
 .theme {
@@ -87,19 +155,19 @@ svg {
   color: var(--text-color);
 }
 
-#theme-toggle {
+/* inputs are hidden, use the input label instead */
+#theme-toggle,
+#menu-button {
   display: none;
-  margin: 0;
-  padding: 0;
-  position: absolute;
 }
 
 /* TODO: */
 /* .active {
   color: var(--highlight-color);
-}
+} */
 
 ::selection {
   background-color: var(--highlight-color);
-  color: var(--text-color);
-} */
+  color: var(--highlight-text-color);
+  opacity: 1;
+}
